Move control script into separate file
