+++ packages/sandbox/container_src/isolation.ts
@@ -28,10 +28,9 @@
 
 import { type ChildProcess, spawn } from 'node:child_process';
 import { randomBytes, randomUUID } from 'node:crypto';
-import { promises as fs } from 'node:fs';
 import * as path from 'node:path';
-import type { ExecEvent, ExecResult } from '../../src/types';
-import type { ProcessRecord, ProcessStatus } from '../types';
+import type { ExecEvent, ExecResult } from '../src/types';
+import type { ProcessRecord, ProcessStatus } from './types';
 
 // Configuration constants
 const CONFIG = {
@@ -68,8 +67,6 @@ interface ControlMessage {
   command?: string;
   cwd?: string;
 }
-
-// Note: buildExecScript helper is now inlined within the generated control script
 
 interface ControlResponse {
   type: 'result' | 'error' | 'ready' | 'stream_event';
@@ -133,20 +130,21 @@ export class Session {
   }
   
   async initialize(): Promise<void> {
-    const controlScript = this.createControlScript();
+    // Use the proper TypeScript control process file
+    const controlProcessPath = path.join(__dirname, 'control-process.js');
     
-    // Write control script to temp file
-    const scriptPath = `${CONFIG.TEMP_DIR}/control_${randomUUID()}.js`;
-    await fs.writeFile(scriptPath, controlScript, 'utf8');
-    
-    // Start control process
-    this.control = spawn('node', [scriptPath], {
+    // Start control process with configuration via environment variables
+    this.control = spawn('node', [controlProcessPath], {
       stdio: ['pipe', 'pipe', 'pipe'],
       env: {
         ...process.env,
         SESSION_ID: this.options.id,
         SESSION_CWD: this.options.cwd || CONFIG.DEFAULT_CWD,
         SESSION_ISOLATED: this.canIsolate ? '1' : '0',
+        COMMAND_TIMEOUT_MS: String(CONFIG.COMMAND_TIMEOUT_MS),
+        CLEANUP_INTERVAL_MS: String(CONFIG.CLEANUP_INTERVAL_MS),
+        TEMP_FILE_MAX_AGE_MS: String(CONFIG.TEMP_FILE_MAX_AGE_MS),
+        TEMP_DIR: CONFIG.TEMP_DIR,
         ...this.options.env
       }
     });
@@ -183,537 +181,9 @@ export class Session {
     // Wait for ready signal
     await this.waitForReady();
     
-    // Clean up temp script
-    await fs.unlink(scriptPath).catch(() => {});
-    
     console.log(`[Session] Session '${this.options.id}' initialized successfully`);
   }
   
-  /**
-   * Generate the control process script that manages the isolated shell.
-   * 
-   * Generated inline rather than using a separate file to:
-   * - Keep everything in one deployable unit
-   * - Inject environment variables at creation time
-   * - Make the implementation self-contained
-   */
-  private createControlScript(): string {
-    return `
-const { spawn } = require('child_process');
-const fs = require('fs');
-const path = require('path');
-const crypto = require('crypto');
-
-// Parse environment
-const sessionId = process.env.SESSION_ID || 'default';
-const sessionCwd = process.env.SESSION_CWD || '${CONFIG.DEFAULT_CWD}';
-const isIsolated = process.env.SESSION_ISOLATED === '1';
-
-// Configuration
-const COMMAND_TIMEOUT_MS = ${CONFIG.COMMAND_TIMEOUT_MS};
-const CLEANUP_INTERVAL_MS = ${CONFIG.CLEANUP_INTERVAL_MS};
-const TEMP_FILE_MAX_AGE_MS = ${CONFIG.TEMP_FILE_MAX_AGE_MS};
-const TEMP_DIR = '${CONFIG.TEMP_DIR}';
-
-console.error(\`[Control] Starting control process for session '\${sessionId}'\`);
-
-// Track active command files for cleanup
-const activeFiles = new Set();
-
-// Cleanup function for orphaned temp files
-function cleanupTempFiles() {
-  try {
-    const files = fs.readdirSync(TEMP_DIR);
-    const now = Date.now();
-    files.forEach(file => {
-      // Match our temp file pattern and check age
-      if (file.match(/^(cmd|out|err|exit)_[a-f0-9-]+/)) {
-        const filePath = path.join(TEMP_DIR, file);
-        try {
-          const stats = fs.statSync(filePath);
-          // Remove files older than max age that aren't active
-          if (now - stats.mtimeMs > TEMP_FILE_MAX_AGE_MS && !activeFiles.has(filePath)) {
-            fs.unlinkSync(filePath);
-            console.error(\`[Control] Cleaned up orphaned temp file: \${file}\`);
-          }
-        } catch (e) {
-          // File might have been removed already
-        }
-      }
-    });
-  } catch (e) {
-    console.error('[Control] Cleanup error:', e);
-  }
-}
-
-// Run cleanup periodically
-setInterval(cleanupTempFiles, CLEANUP_INTERVAL_MS);
-
-// Start the shell with or without isolation
-const shellCommand = isIsolated
-  ? ['unshare', '--pid', '--fork', '--mount-proc', 'bash', '--norc']
-  : ['bash', '--norc'];
-
-const shell = spawn(shellCommand[0], shellCommand.slice(1), {
-  stdio: ['pipe', 'pipe', 'pipe'],
-  cwd: sessionCwd,
-  env: process.env
-});
-
-// Track shell health
-let shellAlive = true;
-
-shell.on('error', (error) => {
-  console.error('[Control] Shell error:', error);
-  shellAlive = false;
-  process.stdout.write(JSON.stringify({
-    type: 'error',
-    id: 'shell',
-    error: error.message
-  }) + '\\n');
-});
-
-shell.on('exit', (code) => {
-  console.error(\`[Control] Shell exited with code \${code}\`);
-  shellAlive = false;
-  // Clean up any remaining temp files
-  activeFiles.forEach(file => {
-    try { fs.unlinkSync(file); } catch (e) {}
-  });
-  process.exit(code || 1);
-});
-
-// Send ready signal
-process.stdout.write(JSON.stringify({ type: 'ready', id: 'init' }) + '\\n');
-
-// Track processing state for each command
-const processingState = new Map();
-
-// Helper function to build exec script with optional cwd handling
-function buildExecScript(cmdFile, outFile, errFile, exitFile, msgId, msgCwd, completionMarker = "DONE") {
-  if (msgCwd) {
-    // If cwd is provided, change directory for this command only
-    // For temporary operations that need specific working directory
-    return \`
-# Execute command with temporary cwd override
-PREV_DIR=$(pwd)
-cd "\${msgCwd}" || { echo "Failed to change directory to \${msgCwd}" > \${errFile}; echo 1 > \${exitFile}; echo "\${completionMarker}:\${msgId}"; return; }
-source \${cmdFile} > \${outFile} 2> \${errFile}
-echo $? > \${exitFile}
-cd "$PREV_DIR"
-echo "\${completionMarker}:\${msgId}"
-\`;
-  } else {
-    // Default behavior - execute in current directory (preserves session state)
-    return \`
-# Execute command in current shell - maintains working directory changes
-source \${cmdFile} > \${outFile} 2> \${errFile}
-echo $? > \${exitFile}
-echo "\${completionMarker}:\${msgId}"
-\`;
-  }
-}
-
-// Handle commands from parent
-process.stdin.on('data', async (data) => {
-  const lines = data.toString().split('\\n');
-  for (const line of lines) {
-    if (!line.trim()) continue;
-    
-    try {
-      const msg = JSON.parse(line);
-      
-      if (msg.type === 'exit') {
-        shell.kill('SIGTERM');
-        process.exit(0);
-      }
-      
-      if (msg.type === 'exec' && msg.command) {
-        if (!shellAlive) {
-          process.stdout.write(JSON.stringify({
-            type: 'error',
-            id: msg.id,
-            error: 'Shell is not alive'
-          }) + '\\n');
-          continue;
-        }
-        
-        // Create temp files for this command
-        const cmdFile = \`\${TEMP_DIR}/cmd_\${msg.id}.sh\`;
-        const outFile = \`\${TEMP_DIR}/out_\${msg.id}\`;
-        const errFile = \`\${TEMP_DIR}/err_\${msg.id}\`;
-        const exitFile = \`\${TEMP_DIR}/exit_\${msg.id}\`;
-        
-        // Track these files as active
-        activeFiles.add(cmdFile);
-        activeFiles.add(outFile);
-        activeFiles.add(errFile);
-        activeFiles.add(exitFile);
-        
-        // Initialize processing state to prevent race conditions
-        processingState.set(msg.id, false);
-        
-        // Write command to file
-        fs.writeFileSync(cmdFile, msg.command, 'utf8');
-        
-        // Use 'source' instead of 'bash' to maintain shell state across commands
-        // - 'bash script.sh' creates a subshell (loses pwd, env vars)
-        // - 'source script.sh' runs in current shell (preserves state)
-        // File-based I/O handles any output type (binary, large, special chars)
-        
-        const execScript = buildExecScript(cmdFile, outFile, errFile, exitFile, msg.id, msg.cwd);
-        
-        shell.stdin.write(execScript);
-        
-        // Set up listener for completion with race condition protection
-        const onData = (chunk) => {
-          const output = chunk.toString();
-          if (output.includes(\`DONE:\${msg.id}\`)) {
-            // Check if already processed (prevents race condition)
-            if (processingState.get(msg.id)) {
-              return;
-            }
-            processingState.set(msg.id, true);
-            
-            // Clear timeout to prevent double cleanup
-            clearTimeout(timeoutId);
-            shell.stdout.off('data', onData);
-            
-            try {
-              // Read results
-              const stdout = fs.readFileSync(outFile, 'utf8');
-              const stderr = fs.readFileSync(errFile, 'utf8');
-              const exitCode = parseInt(fs.readFileSync(exitFile, 'utf8').trim());
-              
-              // Send response
-              process.stdout.write(JSON.stringify({
-                type: 'result',
-                id: msg.id,
-                stdout,
-                stderr,
-                exitCode
-              }) + '\\n');
-              
-              // Cleanup temp files
-              const filesToClean = [cmdFile, outFile, errFile, exitFile];
-              filesToClean.forEach(file => {
-                try {
-                  fs.unlinkSync(file);
-                  activeFiles.delete(file);
-                } catch (e) {
-                  // File might already be deleted
-                }
-              });
-              
-              // Clean up processing state
-              processingState.delete(msg.id);
-            } catch (error) {
-              process.stdout.write(JSON.stringify({
-                type: 'error',
-                id: msg.id,
-                error: \`Failed to read output: \${error.message}\`
-              }) + '\\n');
-              
-              // Still try to clean up files on error
-              [cmdFile, outFile, errFile, exitFile].forEach(file => {
-                try {
-                  fs.unlinkSync(file);
-                  activeFiles.delete(file);
-                } catch (e) {}
-              });
-              processingState.delete(msg.id);
-            }
-          }
-        };
-        
-        // Set up timeout first so we can clear it on success
-        const timeoutId = setTimeout(() => {
-          // Check if already processed
-          if (!processingState.get(msg.id)) {
-            processingState.set(msg.id, true);
-            
-            shell.stdout.off('data', onData);
-            process.stdout.write(JSON.stringify({
-              type: 'error',
-              id: msg.id,
-              error: \`Command timeout after \${COMMAND_TIMEOUT_MS/1000} seconds\`
-            }) + '\\n');
-            
-            // Cleanup files
-            const filesToClean = [cmdFile, outFile, errFile, exitFile];
-            filesToClean.forEach(file => {
-              try {
-                fs.unlinkSync(file);
-                activeFiles.delete(file);
-              } catch (e) {
-                // Silent cleanup - file might already be deleted
-              }
-            });
-            
-            processingState.delete(msg.id);
-          }
-        }, COMMAND_TIMEOUT_MS);
-        
-        shell.stdout.on('data', onData);
-      }
-      
-      if (msg.type === 'exec_stream' && msg.command) {
-        if (!shellAlive) {
-          process.stdout.write(JSON.stringify({
-            type: 'stream_event',
-            id: msg.id,
-            event: {
-              type: 'error',
-              timestamp: new Date().toISOString(),
-              command: msg.command,
-              error: 'Shell is not alive'
-            }
-          }) + '\\n');
-          continue;
-        }
-        
-        // Create temp files for this command
-        const cmdFile = \`\${TEMP_DIR}/cmd_\${msg.id}.sh\`;
-        const outFile = \`\${TEMP_DIR}/out_\${msg.id}\`;
-        const errFile = \`\${TEMP_DIR}/err_\${msg.id}\`;
-        const exitFile = \`\${TEMP_DIR}/exit_\${msg.id}\`;
-        
-        // Track these files as active
-        activeFiles.add(cmdFile);
-        activeFiles.add(outFile);
-        activeFiles.add(errFile);
-        activeFiles.add(exitFile);
-        
-        // Initialize processing state to prevent race conditions
-        processingState.set(msg.id, false);
-        
-        // Write command to file
-        fs.writeFileSync(cmdFile, msg.command, 'utf8');
-        
-        // Start streaming output
-        let stdoutSize = 0;
-        let stderrSize = 0;
-        
-        // Send start event
-        process.stdout.write(JSON.stringify({
-          type: 'stream_event',
-          id: msg.id,
-          event: {
-            type: 'start',
-            timestamp: new Date().toISOString(),
-            command: msg.command
-          }
-        }) + '\\n');
-        
-        // Set up streaming interval to check for new output
-        const streamingInterval = setInterval(() => {
-          try {
-            // Check for new stdout
-            if (fs.existsSync(outFile)) {
-              const currentStdout = fs.readFileSync(outFile, 'utf8');
-              if (currentStdout.length > stdoutSize) {
-                const newData = currentStdout.slice(stdoutSize);
-                stdoutSize = currentStdout.length;
-                process.stdout.write(JSON.stringify({
-                  type: 'stream_event',
-                  id: msg.id,
-                  event: {
-                    type: 'stdout',
-                    timestamp: new Date().toISOString(),
-                    data: newData,
-                    command: msg.command
-                  }
-                }) + '\\n');
-              }
-            }
-            
-            // Check for new stderr
-            if (fs.existsSync(errFile)) {
-              const currentStderr = fs.readFileSync(errFile, 'utf8');
-              if (currentStderr.length > stderrSize) {
-                const newData = currentStderr.slice(stderrSize);
-                stderrSize = currentStderr.length;
-                process.stdout.write(JSON.stringify({
-                  type: 'stream_event',
-                  id: msg.id,
-                  event: {
-                    type: 'stderr',
-                    timestamp: new Date().toISOString(),
-                    data: newData,
-                    command: msg.command
-                  }
-                }) + '\\n');
-              }
-            }
-          } catch (e) {
-            // Files might not exist yet, that's ok
-          }
-        }, 100); // Check every 100ms for real-time feel
-        
-        // Execute command with same pattern as regular exec, including cwd support
-        const execScript = buildExecScript(cmdFile, outFile, errFile, exitFile, msg.id, msg.cwd, "STREAM_DONE");
-        
-        shell.stdin.write(execScript);
-        
-        // Set up listener for completion
-        const onStreamData = (chunk) => {
-          const output = chunk.toString();
-          if (output.includes(\`STREAM_DONE:\${msg.id}\`)) {
-            // Check if already processed (prevents race condition)
-            if (processingState.get(msg.id)) {
-              return;
-            }
-            processingState.set(msg.id, true);
-            
-            // Clear timeout to prevent double cleanup
-            clearTimeout(streamTimeoutId);
-            clearInterval(streamingInterval);
-            shell.stdout.off('data', onStreamData);
-            
-            try {
-              // Read final results
-              const stdout = fs.existsSync(outFile) ? fs.readFileSync(outFile, 'utf8') : '';
-              const stderr = fs.existsSync(errFile) ? fs.readFileSync(errFile, 'utf8') : '';
-              const exitCode = fs.existsSync(exitFile) ? parseInt(fs.readFileSync(exitFile, 'utf8').trim()) : 1;
-              
-              // Send any remaining output
-              if (stdout.length > stdoutSize) {
-                const newData = stdout.slice(stdoutSize);
-                process.stdout.write(JSON.stringify({
-                  type: 'stream_event',
-                  id: msg.id,
-                  event: {
-                    type: 'stdout',
-                    timestamp: new Date().toISOString(),
-                    data: newData,
-                    command: msg.command
-                  }
-                }) + '\\n');
-              }
-              
-              if (stderr.length > stderrSize) {
-                const newData = stderr.slice(stderrSize);
-                process.stdout.write(JSON.stringify({
-                  type: 'stream_event',
-                  id: msg.id,
-                  event: {
-                    type: 'stderr',
-                    timestamp: new Date().toISOString(),
-                    data: newData,
-                    command: msg.command
-                  }
-                }) + '\\n');
-              }
-              
-              // Send completion event
-              process.stdout.write(JSON.stringify({
-                type: 'stream_event',
-                id: msg.id,
-                event: {
-                  type: 'complete',
-                  timestamp: new Date().toISOString(),
-                  command: msg.command,
-                  exitCode: exitCode,
-                  result: {
-                    stdout,
-                    stderr,
-                    exitCode,
-                    success: exitCode === 0
-                  }
-                }
-              }) + '\\n');
-              
-              // Cleanup temp files
-              const filesToClean = [cmdFile, outFile, errFile, exitFile];
-              filesToClean.forEach(file => {
-                try {
-                  fs.unlinkSync(file);
-                  activeFiles.delete(file);
-                } catch (e) {
-                  // File might already be deleted
-                }
-              });
-              
-              // Clean up processing state
-              processingState.delete(msg.id);
-            } catch (error) {
-              process.stdout.write(JSON.stringify({
-                type: 'stream_event',
-                id: msg.id,
-                event: {
-                  type: 'error',
-                  timestamp: new Date().toISOString(),
-                  command: msg.command,
-                  error: \`Failed to read output: \${error.message}\`
-                }
-              }) + '\\n');
-              
-              // Still try to clean up files on error
-              [cmdFile, outFile, errFile, exitFile].forEach(file => {
-                try {
-                  fs.unlinkSync(file);
-                  activeFiles.delete(file);
-                } catch (e) {}
-              });
-              processingState.delete(msg.id);
-            }
-          }
-        };
-        
-        // Set up timeout first so we can clear it on success  
-        const streamTimeoutId = setTimeout(() => {
-          // Check if already processed
-          if (!processingState.get(msg.id)) {
-            processingState.set(msg.id, true);
-            
-            clearInterval(streamingInterval);
-            shell.stdout.off('data', onStreamData);
-            
-            process.stdout.write(JSON.stringify({
-              type: 'stream_event',
-              id: msg.id,
-              event: {
-                type: 'error',
-                timestamp: new Date().toISOString(),
-                command: msg.command,
-                error: \`Command timeout after \${COMMAND_TIMEOUT_MS/1000} seconds\`
-              }
-            }) + '\\n');
-            
-            // Cleanup files
-            const filesToClean = [cmdFile, outFile, errFile, exitFile];
-            filesToClean.forEach(file => {
-              try {
-                fs.unlinkSync(file);
-                activeFiles.delete(file);
-              } catch (e) {
-                // Silent cleanup - file might already be deleted
-              }
-            });
-            
-            processingState.delete(msg.id);
-          }
-        }, COMMAND_TIMEOUT_MS);
-        
-        shell.stdout.on('data', onStreamData);
-      }
-    } catch (e) {
-      console.error('[Control] Failed to parse command:', e);
-    }
-  }
-});
-
-// Cleanup on exit
-process.on('exit', () => {
-  activeFiles.forEach(file => {
-    try { fs.unlinkSync(file); } catch (e) {}
-  });
-});
-
-// Keep process alive
-process.stdin.resume();
-`;
-  }
   
   private waitForReady(): Promise<void> {
     return new Promise((resolve, reject) => {
